requirements-adk.txt
--extra-index-url https://us-python.pkg.dev/vertex-ai/genx-adk/simple/
google-adk
google-genai
fastapi
uvicorn
python-dotenv
